// src/types/custom-d-t.ts
export interface AccordionItem {
  id: string;
  question: string;
  answer: string;
  isExpanded?: boolean;
}

export interface UserStat {
  backgroundColor: string;
  title: string;
  value: string;
  description: string;
  animationDelay: string;
}